requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
mysqlclient==2.2.6
psycopg2-binary==2.9.10
sqlparse==0.5.2
typing_extensions==4.12.2
whitenoise==6.8.2
